feat: Timestamp support
So I perform a new change based on what I did before.

As a proof of being backward compatible, I kept the existing `firestore_test.dart` as is. It sill works but firebase shows warning message to switch to timestamp. There was a bug in dartifyDate that prevented converting js date.

Time to migrate to Timestamp using `firestore.setting(timestampsInSnapshots:true)`!
`firestore_with_timestamp_test.dart` is a copy of `firestore_test.dart` with timestamps enables. No more warning.

A better solution could be found however firebase urge to switch to timestamp and it was currently impossible.